able to query knowledge graph
