+++ data/nlu.yml
@@ -103,18 +103,18 @@ nlu:
     - [tidak pakai]{"entity": "use_public_transport", "value": "tidak pakai kendaraan"}
     - [tidak pakai kendaraan](use_public_transport)
     - [umum]{"entity": "use_public_transport", "value": "kendaraan umum"}
-- intent: schedule_hour
-  examples: |
-    - di [pagi](schedule_hour) hari
-    - berangkat [siang](schedule_hour)
-    - aku berangkat [sore](schedule_hour)
-    - pergi [siang](schedule_hour)
-    - saya pergi di [pagi](schedule_hour) hari
-    - [sore](schedule_hour)
-    - [sore](schedule_hour) hari
-    - [pagi](schedule_hour)
-    - [siang](schedule_hour)
-    - [pagi](schedule_hour)
+# - intent: schedule_hour
+#   examples: |
+#     - di [pagi](schedule_hour) hari
+#     - berangkat [siang](schedule_hour)
+#     - aku berangkat [sore](schedule_hour)
+#     - pergi [siang](schedule_hour)
+#     - saya pergi di [pagi](schedule_hour) hari
+#     - [sore](schedule_hour)
+#     - [sore](schedule_hour) hari
+#     - [pagi](schedule_hour)
+#     - [siang](schedule_hour)
+#     - [pagi](schedule_hour)
 - intent: schedule_day
   examples: |
     - [senin](schedule_day)
