+++ data/stories.yml
@@ -27,12 +27,12 @@ stories:
     - use_public_transport: kendaraan umum
   - slot_was_set:
     - use_public_transport: kendaraan umum
-  - slot_was_set:
-    - requested_slot: schedule_hour
-  - slot_was_set:
-    - schedule_hour: pagi
-  - slot_was_set:
-    - schedule_hour: pagi
+  # - slot_was_set:
+  #   - requested_slot: schedule_hour
+  # - slot_was_set:
+  #   - schedule_hour: pagi
+  # - slot_was_set:
+  #   - schedule_hour: pagi
   - slot_was_set:
     - requested_slot: schedule_day
   - slot_was_set:
@@ -60,5 +60,4 @@ stories:
   - active_loop: null
   - action: utter_slot_values_recommend_form
   - action: utter_submit_recommend_form
-
-
+  - action: action_query_entities
