+++ domain.yml
@@ -3,7 +3,7 @@ session_config:
   session_expiration_time: 60
   carry_over_slots_to_new_session: false
 intents:
-- schedule_hour
+# - schedule_hour
 - use_public_transport
 - schedule_day
 - greet:
@@ -17,10 +17,34 @@ intents:
     use_entities: []
 entities:
 - schedule_day
-- schedule_hour
+# - schedule_hour
 - ticket_price
 - use_public_transport
+- entity_type
+- mention
+- attribute
+- museum
 slots:
+  entity_type:
+    type: text
+    influence_conversation: false
+    # mappings:
+    # - type: custom
+  mention:
+    type: text
+    influence_conversation: false
+    # mappings:
+    # - type: custom
+  attribute:
+    type: text
+    influence_conversation: false
+    # mappings:
+    # - type: custom
+  museum:
+    type: text
+    influence_conversation: false
+    # mappings:
+    # - type: custom
   use_public_transport:
     type: categorical
     influence_conversation: false
@@ -28,13 +52,13 @@ slots:
     - kendaraan umum
     - kendaraan pribadi
     - tidak pakai kendaraan
-  schedule_hour:
-    type: categorical
-    influence_conversation: false
-    values:
-    - pagi
-    - siang
-    - sore
+  # schedule_hour:
+  #   type: categorical
+  #   influence_conversation: false
+  #   values:
+  #   - pagi
+  #   - siang
+  #   - sore
   schedule_day:
     type: list
     influence_conversation: false
@@ -55,8 +79,8 @@ responses:
   - text: Proses rekomendasi dibatalkan
   utter_ask_use_public_transport:
   - text: Apakah kamu akan menggunakan kendaraan umum, kendaraan pribadi atau tidak pakai kendaraan?
-  utter_ask_schedule_hour:
-  - text: Berangkat pagi, siang, sore?
+  # utter_ask_schedule_hour:
+  # - text: Berangkat pagi, siang, sore?
   utter_ask_schedule_day:
   - text: |-
       Di hari apa saja kamu bisa mengunjungi museum? 
@@ -69,10 +93,13 @@ responses:
   utter_submit_recommend_form:
   - text: Mohon tunggu sebentar, Aku sedang memproses rekomendasi museum untuk kamu...
   utter_slot_values_recommend_form:
-  - text: Kamu menjawab {use_public_transport}, {schedule_hour}, {schedule_day}, {ticket_price}
+  - text: Kamu menjawab {use_public_transport}, {schedule_day}, {ticket_price}
+  utter_ask_rephrase:
+  - text: "Sorry, I'm not sure I understand. Can you rephrase?"
+  - text: "Can you please rephrase? I did not got that."
 actions:
 - utter_ask_schedule_day
-- utter_ask_schedule_hour
+# - utter_ask_schedule_hour
 - utter_ask_ticket_price
 - utter_ask_use_public_transport
 - utter_cancel_recommendation
@@ -81,6 +108,8 @@ actions:
 - utter_slot_values_recommend_form
 - utter_submit_recommend_form
 - utter_user_agree_recommend
+- utter_ask_rephrase
+- action_query_entities
 forms:
   recommend_form:
     required_slots:
@@ -88,10 +117,10 @@ forms:
       - type: from_entity
         entity: use_public_transport
         intent: use_public_transport
-      schedule_hour:
-      - type: from_entity
-        entity: schedule_hour
-        intent: schedule_hour
+      # schedule_hour:
+      # - type: from_entity
+      #   entity: schedule_hour
+      #   intent: schedule_hour
       schedule_day:
       - type: from_entity
         entity: schedule_day
