--- requirements.txt
@@ -1,157 +1,0 @@
-absl-py==0.13.0
-aio-pika==6.8.0
-aiofiles==0.7.0
-aiohttp==3.7.4
-aiormq==3.3.1
-aiosignal==1.2.0
-alembic==1.4.3
-APScheduler==3.7.0
-astunparse==1.6.3
-async-generator==1.10
-async-timeout==3.0.1
-attrs==19.3.0
-backports.zoneinfo==0.2.1
-bidict==0.21.4
-boto3==1.20.5
-botocore==1.23.5
-CacheControl==0.12.10
-cachetools==4.2.4
-certifi==2021.10.8
-cffi==1.15.0
-chardet==3.0.4
-charset-normalizer==2.0.7
-clang==5.0
-cloudpickle==1.6.0
-colorclass==2.2.0
-coloredlogs==15.0.1
-colorhash==1.0.3
-croniter==1.0.15
-cryptography==3.4.8
-cycler==0.11.0
-decorator==4.4.2
-dm-tree==0.1.6
-dnspython==1.16.0
-docopt==0.6.2
-fbmessenger==6.0.0
-flatbuffers==1.12
-frozenlist==1.2.0
-future==0.18.2
-gast==0.4.0
-gitdb==4.0.9
-GitPython==3.1.24
-google-auth==2.3.3
-google-auth-oauthlib==0.4.6
-google-pasta==0.2.0
-greenlet==1.1.2
-grpcio==1.38.0
-h11==0.9.0
-h5py==3.1.0
-httpcore==0.11.1
-httptools==0.3.0
-httpx==0.15.4
-humanfriendly==10.0
-idna==3.3
-isodate==0.6.0
-jmespath==0.10.0
-joblib==1.0.1
-jsonpickle==2.0.0
-jsonschema==3.2.0
-kafka-python==1.4.7
-keras==2.6.0
-Keras-Preprocessing==1.1.2
-kiwisolver==1.3.2
-Mako==1.1.5
-Markdown==3.3.4
-MarkupSafe==2.0.1
-matplotlib==3.3.4
-mattermostwrapper==2.2
-msgpack==1.0.2
-multidict==5.0.0
-networkx==2.5.1
-numpy==1.19.5
-oauthlib==3.1.1
-opt-einsum==3.3.0
-packaging==20.9
-pamqp==2.3.0
-pika==1.2.0
-Pillow==8.4.0
-pkg_resources==0.0.0
-prompt-toolkit==2.0.10
-protobuf==3.15.5
-psycopg2-binary==2.9.2
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-pycparser==2.21
-pydot==1.4.2
-PyJWT==2.0.0
-pykwalify==1.8.0
-pymongo==3.10.1
-pyparsing==3.0.6
-pyrsistent==0.18.0
-pyTelegramBotAPI==3.8.3
-python-crfsuite==0.9.7
-python-dateutil==2.8.2
-python-editor==1.0.4
-python-engineio==4.3.0
-python-socketio==5.4.1
-pytz==2021.3
-pytz-deprecation-shim==0.1.0.post0
-questionary==1.5.2
-rasa==2.8.13
-rasa-sdk==2.8.2
-rasa-x==0.42.5
-redis==3.5.3
-regex==2021.7.6
-requests==2.26.0
-requests-oauthlib==1.3.0
-requests-toolbelt==0.9.1
-rfc3986==1.5.0
-rocketchat-API==1.16.0
-rsa==4.7.2
-ruamel.yaml==0.16.13
-ruamel.yaml.clib==0.2.6
-s3transfer==0.5.0
-sanic==20.9.1
-Sanic-Cors==0.10.0.post3
-sanic-jwt==1.6.0
-Sanic-Plugins-Framework==0.9.5
-scikit-learn==0.24.2
-scipy==1.7.2
-sentry-sdk==1.2.0
-six==1.16.0
-sklearn-crfsuite==0.3.6
-slackclient==2.9.3
-smmap==5.0.0
-sniffio==1.2.0
-SQLAlchemy==1.4.27
-tabulate==0.8.9
-tarsafe==0.0.3
-tensorboard==2.6.0
-tensorboard-data-server==0.6.1
-tensorboard-plugin-wit==1.8.0
-tensorflow==2.6.2
-tensorflow-addons==0.14.0
-tensorflow-estimator==2.6.0
-tensorflow-hub==0.12.0
-tensorflow-probability==0.13.0
-tensorflow-text==2.6.0
-termcolor==1.1.0
-terminaltables==3.1.0
-threadpoolctl==3.0.0
-tqdm==4.62.3
-twilio==6.50.1
-typedb-client==2.5.0
-typedb-protocol==2.5.0
-typeguard==2.13.0
-typing-extensions==3.10.0.2
-tzdata==2021.5
-tzlocal==2.1
-ujson==1.35
-urllib3==1.26.7
-uvloop==0.16.0
-wcwidth==0.2.5
-webexteamssdk==1.6
-websockets==8.1
-Werkzeug==2.0.2
-wrapt==1.12.1
-yarl==1.7.2
