+++ story_graph.dot
@@ -5,25 +5,25 @@ digraph  {
 2 [class=active, fontsize=12, label=utter_greet];
 3 [class=active, fontsize=12, label=utter_user_agree_recommend];
 4 [class=active, fontsize=12, label=recommend_form];
+8 [class=active, fontsize=12, label=utter_slot_values_recommend_form];
 9 [class=active, fontsize=12, label=utter_submit_recommend_form];
-10 [class=active, fontsize=12, label=utter_slot_values_recommend_form];
+10 [class=active, fontsize=12, label=action_query_entities];
 11 [class="intent dashed active", label="  ?  ", shape=rect];
-12 [class="intent active", fillcolor=lightblue, label=hi, shape=rect, style=filled];
-13 [class="intent active", fillcolor=lightblue, label="ya mau dong", shape=rect, style=filled];
+12 [class="intent active", fillcolor=lightblue, label=helo, shape=rect, style=filled];
+13 [class="intent active", fillcolor=lightblue, label=mau, shape=rect, style=filled];
 14 [class="intent active", fillcolor=lightblue, label=umum, shape=rect, style=filled];
-15 [class="intent active", fillcolor=lightblue, label=pagi, shape=rect, style=filled];
-16 [class="intent active", fillcolor=lightblue, label="senin minggu jumat", shape=rect, style=filled];
-17 [class="intent active", fillcolor=lightblue, label="123 90859", shape=rect, style=filled];
+15 [class="intent active", fillcolor=lightblue, label="senin selasa", shape=rect, style=filled];
+16 [class="intent active", fillcolor=lightblue, label="0 50000", shape=rect, style=filled];
 0 -> "-1"  [class="", key=NONE, label=""];
 0 -> 1  [class=active, key=NONE, label=""];
 1 -> 12  [class=active, key=0];
 2 -> 13  [class=active, key=0];
 3 -> 4  [class=active, key=NONE, label=""];
-4 -> 9  [class=active, key=NONE, label=""];
+4 -> 8  [class=active, key=NONE, label=""];
 4 -> 14  [class=active, key=0];
 4 -> 15  [class=active, key=0];
 4 -> 16  [class=active, key=0];
-4 -> 17  [class=active, key=0];
+8 -> 9  [class=active, key=NONE, label=""];
 9 -> 10  [class=active, key=NONE, label=""];
 10 -> 11  [class=active, key=NONE, label=""];
 12 -> 2  [class=active, key=0];
@@ -31,5 +31,4 @@ digraph  {
 14 -> 4  [class=active, key=0];
 15 -> 4  [class=active, key=0];
 16 -> 4  [class=active, key=0];
-17 -> 4  [class=active, key=0];
 }
